fixup

+++ playground/src/proxies/hwui/panel-unit/boled/parameter-screens/controls/MiniParameterLabel.cpp
@@ -62,10 +62,10 @@ void MiniParameterLabel::setFontColor(FrameBuffer &fb) const
     fb.setColor(FrameBuffer::Colors::C128);
 }
 
-const std::pair<Glib::ustring, short> MiniParameterLabel::prepareDisplayString(const Parameter *p) const
+const std::pair<Glib::ustring, size_t> MiniParameterLabel::prepareDisplayString(const Parameter *p) const
 {
   auto changed = p->isChangedFromLoaded();
   auto text = p->getMiniParameterEditorName() + (changed ? "*" : "");
-  short suffixLen = changed ? 1 : 0;
+  size_t suffixLen = changed ? 1 : 0;
   return std::make_pair(text, suffixLen);
 }
